Add AddCollection component for managing post collections and integrate with Post component
